color sensor work in progress

+++ src/ca/mcgill/ecse211/project/sensor/ColorUser.java
@@ -4,6 +4,6 @@ import lejos.robotics.SampleProvider;
 
 public interface ColorUser {
   
-  public void processColorData(int color);
+  public void processColorData(float color);
   
 }
